requirements.txt
streamlit
langchain
openai
sqlalchemy
pyodbc
python-dotenv
